worker/requirements.txt
pika
requests
openai-whisper
